Added additional services to test in card and pba
