+++ api/src/functionalTest/java/uk/gov/hmcts/payment/functional/fixture/PaymentFixture.java
@@ -55,6 +55,26 @@ public class PaymentFixture {
             .build();
     }
 
+    public static CardPaymentRequest cardPaymentRequestall(String amountString, Service service) {
+        String ccdCaseNumber = "1111-CC12-" + RandomUtils.nextInt();
+        return CardPaymentRequest.createCardPaymentRequestDtoWith()
+            .amount(new BigDecimal(amountString))
+            .description("description")
+            .caseReference("telRefNumber")
+            .ccdCaseNumber(ccdCaseNumber)
+            .service(service)
+            .siteId("string")
+            .currency(CurrencyCode.GBP)
+            .fees(Lists.newArrayList(
+                FeeDto.feeDtoWith()
+                    .calculatedAmount(new BigDecimal(amountString))
+                    .code("FEE0002")
+                    .version("1")
+                    .build())
+            )
+            .build();
+    }
+
     public static CreditAccountPaymentRequest aPbaPaymentRequest(String amountString, Service service) {
         String ccdCaseNumber = "1111-CC12-" + RandomUtils.nextInt();
         return CreditAccountPaymentRequest.createCreditAccountPaymentRequestDtoWith()
@@ -95,6 +115,29 @@ public class PaymentFixture {
                 FeeDto.feeDtoWith()
                     .calculatedAmount(new BigDecimal(amountString))
                     .code("FEE0001")
+                    .version("1")
+                    .build())
+            )
+            .build();
+    }
+
+    public static CreditAccountPaymentRequest aPbaPaymentRequestForDivorce(String amountString, Service service) {
+        String ccdCaseNumber = "1111-CC12-" + RandomUtils.nextInt();
+        return CreditAccountPaymentRequest.createCreditAccountPaymentRequestDtoWith()
+            .amount(new BigDecimal(amountString))
+            .description("New passport application")
+            .ccdCaseNumber(ccdCaseNumber)
+            .caseReference("aCaseReference")
+            .service(service)
+            .currency(CurrencyCode.GBP)
+            .siteId("AA08")
+            .customerReference("CUST101")
+            .organisationName("ORG101")
+            .accountNumber("PBAFUNC12345")
+            .fees(Lists.newArrayList(
+                FeeDto.feeDtoWith()
+                    .calculatedAmount(new BigDecimal(amountString))
+                    .code("FEE0002")
                     .version("1")
                     .build())
             )
